Provide a parent class for tsdb and graphite since they use the same functionality
